Correccion del error al renderizar y se separa drawer y speed dial en componentes diferentes
